+++ apps/elf-eval/src/bin/live_baseline_elf/providers/deterministic.rs
@@ -1,0 +1,66 @@
+use crate::{
+	Arc, BoxFuture, EmbeddingProvider, EmbeddingProviderConfig, ExtractorProvider,
+	LlmProviderConfig, ProviderConfig, Providers, RerankProvider, Value,
+};
+use elf_service::Result;
+
+#[derive(Debug)]
+struct DeterministicEmbedding {
+	vector_dim: u32,
+}
+impl EmbeddingProvider for DeterministicEmbedding {
+	fn embed<'a>(
+		&'a self,
+		_cfg: &'a EmbeddingProviderConfig,
+		texts: &'a [String],
+	) -> BoxFuture<'a, Result<Vec<Vec<f32>>>> {
+		let dim = self.vector_dim;
+		let vectors = texts.iter().map(|text| crate::embed_text(text, dim)).collect();
+
+		Box::pin(async move { Ok(vectors) })
+	}
+}
+
+#[derive(Debug)]
+struct TokenOverlapRerank;
+impl RerankProvider for TokenOverlapRerank {
+	fn rerank<'a>(
+		&'a self,
+		_cfg: &'a ProviderConfig,
+		query: &'a str,
+		docs: &'a [String],
+	) -> BoxFuture<'a, Result<Vec<f32>>> {
+		let query_terms = crate::terms(query);
+		let scores = docs
+			.iter()
+			.map(|doc| {
+				let doc_terms = crate::terms(doc);
+				let hits = query_terms.intersection(&doc_terms).count() as f32;
+
+				hits / query_terms.len().max(1) as f32
+			})
+			.collect();
+
+		Box::pin(async move { Ok(scores) })
+	}
+}
+
+#[derive(Debug)]
+struct NoopExtractor;
+impl ExtractorProvider for NoopExtractor {
+	fn extract<'a>(
+		&'a self,
+		_cfg: &'a LlmProviderConfig,
+		_messages: &'a [Value],
+	) -> BoxFuture<'a, Result<Value>> {
+		Box::pin(async move { Ok(serde_json::json!({ "notes": [] })) })
+	}
+}
+
+pub(crate) fn deterministic_providers(vector_dim: u32) -> Providers {
+	Providers::new(
+		Arc::new(DeterministicEmbedding { vector_dim }),
+		Arc::new(TokenOverlapRerank),
+		Arc::new(NoopExtractor),
+	)
+}
